Update README.md
I have made changes in the file without losing any content. The purpose was to enhance readability, and keep the content concise. 
Some longer sentences are switched with concise words, heading have been made bold for highlighting.  
(This is a part of an assignment we received from our university)
Thank you!